geolocation service fix circular import minor
