+++ ZelBack/src/services/serviceManager.js
@@ -8,6 +8,7 @@ const fluxNetworkHelper = require('./fluxNetworkHelper');
 const appsService = require('./appsService');
 const daemonServiceMiscRpcs = require('./daemonService/daemonServiceMiscRpcs');
 const fluxService = require('./fluxService');
+const geolocationService = require('./geolocationService');
 const upnpService = require('./upnpService');
 const userconfig = require('../../../config/userconfig');
 
@@ -76,7 +77,7 @@ async function startFluxFunctions() {
     log.info('Connections polling prepared');
     daemonServiceMiscRpcs.daemonBlockchainInfoService();
     log.info('Flux Daemon Info Service Started');
-    fluxService.InstallFluxWatchTower();
+    fluxService.installFluxWatchTower();
     fluxNetworkHelper.checkDeterministicNodesCollisions();
     log.info('Flux checks operational');
     fluxCommunication.fluxDiscovery();
@@ -93,7 +94,7 @@ async function startFluxFunctions() {
     }, 1 * 60 * 1000);
     setTimeout(() => {
       log.info('Starting setting Node Geolocation');
-      fluxService.setNodeGeolocation();
+      geolocationService.setNodeGeolocation();
     }, 90 * 1000);
     setTimeout(() => { // wait as of restarts due to ui building
       explorerService.initiateBlockProcessor(true, true);
