+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "flux",
-  "version": "3.20.0",
+  "version": "3.21.0",
   "description": "Flux, Your Gateway to a Decentralized World",
   "repository": {
     "type": "git",
